bdf:  - more DMNCON support  - CFAST conversion op2 geom:  - more DMNCON, FLUTTER, NLRSFD support  - MAT2 64-bit support  - improved 64-bit PCOMP support (no -1 flags for single PCOMP) op2 writer:  - MSC SPC bug fixed  - testing PFAST  - more PBUSH support

+++ pyNastran/bdf/cards/properties/properties.py
@@ -70,7 +70,7 @@ class PFAST(Property):
         ----------
         pid : int
             property id
-        d : int
+        d : float
             diameter of the fastener
         kt1, kt2, kt3 : float
             stiffness values in directions 1-3
